Ajout fichiers personnels Ajout modification des temps et des preferences Ajout affichage arretes dans les chemins

+++ Fonctions.h
@@ -5,7 +5,7 @@
 #include "Sommet.h"
 
 
-float calculTps(std::string type, Sommet* s1, Sommet* s2, std::vector<std::pair<std::string, std::pair<int, int>>> tabRemontees, std::vector<std::pair<std::string, int>> tabDescentes);
+float calculTps(std::string type, Sommet* s1, Sommet* s2, std::vector<std::pair<std::string, std::pair<float, float>>> tabRemontees, std::vector<std::pair<std::string, float>> tabDescentes);
 
 void menu();
